Changes to be committed: 	modified:   src/main/java/it/unitn/ds1/Client.java 	modified:   src/main/java/it/unitn/ds1/Replica.java 	modified:   src/test/java/it/unitn/ds1/ElectionTest.java Fixed merge conflicts
It remains to clean up some things for example we should drop either
`tellWithDelay` or `sendDelayes` since they're absolutely equivalent implementations
of the same thing.

Also:
* I don't quite get why you wanted to move `writeIndex++` from `onUpdateRequestMsg`
to `onWriteMsg`. I tried doing so, but that breakes crash detection for some reason,
so I put it back. If moving it really makes sense we should investigate why crash
detection breaks.
* Regarding the use of `WriteId` in place of `SimpleEntry`, I've already changed all
what was to be changed in my branch, so the greates part of the hassle is done. The
only thing left would be applying the same changes in the election part. I see clear
advantages in code clarity with the use of `WriteId`.